refactor(whatsapp-stay): isolate Statsig into its own reusable module (FOU-550)
Extract the Statsig client lifecycle (lazy init + fail-closed + best-effort)
into src/misc/statsig.ts exposing a generic isGateEnabled(gateName, userID, logger),
so future gates can reuse it. feature-toggle.ts becomes a thin consumer keeping
the WhatsApp gate constant and the phone -> userID mapping.
